requirements.txt
lxml >= 3.5.0, < 5
coveralls
Jinja2
signxml
urllib3 >= 1.22
suds-jurko >= 0.6
suds-jurko-requests >= 1.1
defusedxml >= 0.4.1, < 1
eight >= 0.3.0, < 1
cryptography >= 1.8, < 3
pyOpenSSL >= 16.0.0, < 18
certifi >= 2015.11.20.1
xmlsec >= 1.3.3
reportlab
pytest>=4.1.1
pytest-cov
pytz
zeep
